Move query and data structures into Agent-specific file
