+++ thoth-app/src/agent/mod.rs
@@ -1,3 +1,4 @@
 pub mod notification_bus;
 pub mod session_timer;
 pub mod contributor_links;
+pub mod contributor_links_query;
