+++ thoth-app/src/models/contributor/contributor_query.rs
@@ -12,16 +12,6 @@ pub const CONTRIBUTOR_QUERY: &str = "
             fullName
             orcid
             website
-            contributions {
-                work {
-                    title
-                    imprint {
-                        publisher {
-                            publisherName
-                        }
-                    }
-                }
-            }
         }
     }
 ";
